chore: improved key binding
