+++ site/src/components/Layout.tsx
@@ -18,9 +18,10 @@ import Logo from './Logo';
 
 interface LayoutProps extends BoxProps {
   aside?: BoxProps['children'];
+  keyBindings?: Parameters<typeof useKeyBindings>[0];
 }
 
-const Layout = ({ aside, children }: LayoutProps) => {
+const Layout = ({ aside, children, keyBindings }: LayoutProps) => {
   const router = useRouter();
   const { toggleMobileMenu } = useMobileNavigationContext();
   const { toggleColorMode } = useColorMode();
@@ -37,22 +38,26 @@ const Layout = ({ aside, children }: LayoutProps) => {
   };
 
   function setQuery(query) {
-    router.push({
+    router.push(
+      {
         pathname: '/',
         query: { query: query },
-    },
-    undefined,
-    { shallow: true })
+      },
+      undefined,
+      { shallow: true }
+    );
   }
 
-  useKeyBindings({
-    Escape: {
-      fn: () => setQuery(''),
-    },
-    KeyT: {
-      fn: () => toggleColorMode(),
-    },
-  });
+  useKeyBindings(
+    keyBindings || {
+      Escape: {
+        fn: () => setQuery(''),
+      },
+      t: {
+        fn: () => toggleColorMode(),
+      },
+    }
+  );
 
   return (
     <Box h="100vh">
